requirements.txt
tqdm
pillow
pillow-avif-plugin
